+++ ShipCamPro/bundle/CHANGELOG.md
@@ -1,5 +1,11 @@
 ### Changelog
 
+##### 0.3.0
+- Removed the (now) unnecessary camera correction
+- Added zoom (mousewheel)
+- Added 360 degree camera
+- Added config settings
+
 ##### 0.2.0
 - The 3rd person freecam now stays in place